move sliding_window test to correct script
